requirements.txt
python_dotenv
streamlit
google-generativeai
